feat: add: 'Display the current page in the browser URL using query parameters'
