+++ components/index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>React RSS App</title>
+  <title>APP</title>
   <link rel="icon" href="https://cdn-icons-png.flaticon.com/512/3336/3336008.png" />
 </head>
